MM-44718 finish run extacted to hook  (#1262)
Extract finish run behavior to a hook to share among context menu and finish run section.

Fixes https://mattermost.atlassian.net/browse/MM-44718
